playground default for qir gen is adaptive_rif and looks for entrypoint target profile
